Remove potential cycle between OptionsInitializer and EngineInitializer by moving computed options.
[ci skip-rust]

[ci skip-build-wheels]
