+++ src/python/pants/init/engine_initializer.py
@@ -30,9 +30,8 @@ from pants.engine.streaming_workunit_handler import rules as streaming_workunit_
 from pants.engine.target import RegisteredTargetTypes
 from pants.engine.unions import UnionMembership
 from pants.init import specs_calculator
-from pants.init.options_initializer import OptionsInitializer
-from pants.option.global_options import DEFAULT_EXECUTION_OPTIONS, ExecutionOptions
-from pants.option.options import Options
+from pants.option.global_options import DEFAULT_EXECUTION_OPTIONS, ExecutionOptions, GlobalOptions
+from pants.option.options_bootstrapper import OptionsBootstrapper
 from pants.option.subsystem import Subsystem
 from pants.util.ordered_set import FrozenOrderedSet
 from pants.vcs.changed import rules as changed_rules
@@ -164,25 +163,24 @@ class EngineInitializer:
 
     @staticmethod
     def setup_graph(
-        options: Options,
+        options_bootstrapper: OptionsBootstrapper,
         build_configuration: BuildConfiguration,
         executor: Optional[PyExecutor] = None,
     ) -> GraphScheduler:
         native = Native()
         build_root = get_buildroot()
-        bootstrap_options = options.bootstrap_option_values()
+        bootstrap_options = options_bootstrapper.bootstrap_options.for_global_scope()
+        options = options_bootstrapper.full_options(build_configuration)
         assert bootstrap_options is not None
         executor = executor or PyExecutor(
-            *OptionsInitializer.compute_executor_arguments(bootstrap_options)
+            *GlobalOptions.compute_executor_arguments(bootstrap_options)
         )
         return EngineInitializer.setup_graph_extended(
             build_configuration,
             ExecutionOptions.from_options(options),
             native=native,
             executor=executor,
-            pants_ignore_patterns=OptionsInitializer.compute_pants_ignore(
-                build_root, bootstrap_options
-            ),
+            pants_ignore_patterns=GlobalOptions.compute_pants_ignore(build_root, bootstrap_options),
             use_gitignore=bootstrap_options.pants_ignore_use_gitignore,
             local_store_dir=bootstrap_options.local_store_dir,
             local_execution_root_dir=bootstrap_options.local_execution_root_dir,
